Add 'type' parameter to pointedBlockChain function
If type == 1, each clique has a single "point" connected to all the
other points; thus, the points form a k-clique.

If type == 2, each clique has k - 1 points, each of which is connected
to exactly one point in one of the other cliques; thus, the points form
a perfect matching on k(k - 1) vertices.